FA-287 Disabled dragging password inputs
There's an issue with angular-drag-and-drop-lists with password fields where it will start the drag operation but never call drag end or anything to stop the operation. Likely a security measure for passwords. This gets the dndDragging/dndDraggingSource classes stuck which hide the component. I added dnd-nodrag to password fields so they can't be dragged.